Change the namespace of the Attributes classes
